count neopixel delay in transitiontime + platformIO projects update
