+++ Lights/Arduino/SK6812HueStrip/SK6812HueStrip.ino
@@ -283,6 +283,7 @@ void setup() {
 
   server.on("/switch", []() {
     server.send(200, "text/plain", "OK");
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
     int button;
     for (uint8_t i = 0; i < server.args(); i++) {
       if (server.argName(i) == "button") {
@@ -325,9 +326,9 @@ void setup() {
       }
       for (uint8_t j = 0; j <= 3; j++) {
         if (light_state[i]) {
-          step_level[i][j] = (rgbw[i][j] - current_rgbw[i][j]) / 54;
+          step_level[i][j] = ((float)rgbw[i][j] - current_rgbw[i][j]) / transitiontime;
         } else {
-          step_level[i][j] = current_rgbw[i][j] / 54;
+          step_level[i][j] = current_rgbw[i][j] / transitiontime;
         }
       }
     }
@@ -412,7 +413,7 @@ void setup() {
         transitiontime = server.arg(i).toInt();
       }
     }
-    transitiontime *= 10;
+    transitiontime *= 10 - (pixelCount / 40); //every extra led add a small delay that need to be counted
     server.send(200, "text/plain", "OK, x: " + (String)x[light] + ", y:" + (String)y[light] + ", bri:" + (String)bri[light] + ", ct:" + ct[light] + ", colormode:" + color_mode[light] + ", state:" + light_state[light]);
     if (color_mode[light] == 1 && light_state[light] == true) {
       convert_xy(light);
@@ -445,7 +446,7 @@ void setup() {
   });
 
   server.on("/", []() {
-    float transitiontime = 20;
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
     if (server.hasArg("startup")) {
       if (  EEPROM.read(1) != server.arg("startup").toInt()) {
         EEPROM.write(1, server.arg("startup").toInt());
