+++ Lights/Arduino/WS2812BHueStrip/WS2812BHueStrip.ino
@@ -7,8 +7,8 @@
 #include <WiFiManager.h>
 #include <EEPROM.h>
 
-#define lightsCount 3
-#define pixelCount 60
+#define lightsCount 4
+#define pixelCount 300
 
 // if you want to setup static ip uncomment these 3 lines and line 72
 //IPAddress strip_ip ( 192,  168,   10,  95);
@@ -282,6 +282,7 @@ void setup() {
 
   server.on("/switch", []() {
     server.send(200, "text/plain", "OK");
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
     int button;
     for (uint8_t i = 0; i < server.args(); i++) {
       if (server.argName(i) == "button") {
@@ -324,9 +325,9 @@ void setup() {
       }
       for (uint8_t j = 0; j < 3; j++) {
         if (light_state[i]) {
-          step_level[i][j] = (rgb[i][j] - current_rgb[i][j]) / 54;
+          step_level[i][j] = ((float)rgb[i][j] - current_rgb[i][j]) / transitiontime;
         } else {
-          step_level[i][j] = current_rgb[i][j] / 54;
+          step_level[i][j] = current_rgb[i][j] / transitiontime;
         }
       }
     }
@@ -407,7 +408,7 @@ void setup() {
         transitiontime = server.arg(i).toInt();
       }
     }
-    transitiontime *= 10;
+    transitiontime *= 10 - (pixelCount / 40); //every extra led add a small delay that need to be counted
     server.send(200, "text/plain", "OK, x: " + (String)x[light] + ", y:" + (String)y[light] + ", bri:" + (String)bri[light] + ", ct:" + ct[light] + ", colormode:" + color_mode[light] + ", state:" + light_state[light]);
     if (color_mode[light] == 1 && light_state[light] == true) {
       convert_xy(light);
@@ -440,7 +441,7 @@ void setup() {
   });
 
   server.on("/", []() {
-    float transitiontime = 20;
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
     if (server.hasArg("startup")) {
       if (  EEPROM.read(1) != server.arg("startup").toInt()) {
         EEPROM.write(1, server.arg("startup").toInt());
@@ -629,7 +630,7 @@ void lightEngine() {
         strip.Show();
       }
     }
-    delay(2);
+    delay(1);
   }
 }
