+++ Lights/PlatformIO/GenericWifiHueLight/.travis.yml
