+++ Lights/PlatformIO/GenericWifiHueLight/platformio.ini
@@ -1,0 +1,37 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; http://docs.platformio.org/page/projectconf.html
+
+[common]
+lib_deps_external =
+  WiFiManager
+
+[env:nodemcuv2]
+platform = espressif8266
+framework = arduino
+board = nodemcuv2
+build_flags = -Wl,-Tesp8266.flash.4m1m.ld -D FIRMWARE_VARIANT=nodemcuv2
+lib_deps =
+  ${common.lib_deps_external}
+
+[env:d1_mini]
+platform = espressif8266
+framework = arduino
+board = d1_mini
+build_flags = -Wl,-Tesp8266.flash.4m1m.ld -D FIRMWARE_VARIANT=d1_mini
+lib_deps =
+  ${common.lib_deps_external}
+
+[env:esp12]
+platform = espressif8266
+board = esp12e
+framework = arduino
+build_flags = -Wl,-Tesp8266.flash.4m1m.ld -D FIRMWARE_VARIANT=esp12
+lib_deps =
+  ${common.lib_deps_external}
