+++ Lights/PlatformIO/GenericWifiHueLight/src/GenericWifiHueLight.cpp
@@ -1,0 +1,610 @@
+#include <Arduino.h>
+
+#include <ESP8266WiFi.h>
+#include <ESP8266mDNS.h>
+#include <WiFiUdp.h>
+#include <ArduinoOTA.h>
+#include <ESP8266WebServer.h>
+#include <WiFiManager.h>
+#include <EEPROM.h>
+
+
+
+#define red_pin 12
+#define green_pin 13
+#define blue_pin 14
+#define white_pin 5
+
+#define pwm_channels 4 // put 3 for rgb, 4 for rgbw
+
+// if you want to setup static ip uncomment these 3 lines and line 72
+//IPAddress strip_ip ( 192,  168,   10,  95);
+//IPAddress gateway_ip ( 192,  168,   10,   1);
+//IPAddress subnet_mask(255, 255, 255,   0);
+
+uint8_t rgbw[4], color_mode, scene, pins[4];
+bool light_state;
+int transitiontime, ct, hue, bri, sat;
+float step_level[4], current_rgbw[4], x, y;
+byte mac[6];
+
+ESP8266WebServer server(80);
+
+void convert_hue()
+{
+  double      hh, p, q, t, ff, s, v;
+  long        i;
+
+  rgbw[3] = 0;
+  s = sat / 255.0;
+  v = bri / 255.0;
+
+  if (s <= 0.0) {      // < is bogus, just shuts up warnings
+    rgbw[0] = v;
+    rgbw[1] = v;
+    rgbw[2] = v;
+    return;
+  }
+  hh = hue;
+  if (hh >= 65535.0) hh = 0.0;
+  hh /= 11850, 0;
+  i = (long)hh;
+  ff = hh - i;
+  p = v * (1.0 - s);
+  q = v * (1.0 - (s * ff));
+  t = v * (1.0 - (s * (1.0 - ff)));
+
+  switch (i) {
+    case 0:
+      rgbw[0] = v * 255.0;
+      rgbw[1] = t * 255.0;
+      rgbw[2] = p * 255.0;
+      break;
+    case 1:
+      rgbw[0] = q * 255.0;
+      rgbw[1] = v * 255.0;
+      rgbw[2] = p * 255.0;
+      break;
+    case 2:
+      rgbw[0] = p * 255.0;
+      rgbw[1] = v * 255.0;
+      rgbw[2] = t * 255.0;
+      break;
+
+    case 3:
+      rgbw[0] = p * 255.0;
+      rgbw[1] = q * 255.0;
+      rgbw[2] = v * 255.0;
+      break;
+    case 4:
+      rgbw[0] = t * 255.0;
+      rgbw[1] = p * 255.0;
+      rgbw[2] = v * 255.0;
+      break;
+    case 5:
+    default:
+      rgbw[0] = v * 255.0;
+      rgbw[1] = p * 255.0;
+      rgbw[2] = q * 255.0;
+      break;
+  }
+
+}
+
+void convert_xy()
+{
+  float Y = bri / 250.0f;
+
+  float z = 1.0f - x - y;
+
+  float X = (Y / y) * x;
+  float Z = (Y / y) * z;
+
+  // sRGB D65 conversion
+  float r =  X * 1.656492f - Y * 0.354851f - Z * 0.255038f;
+  float g = -X * 0.707196f + Y * 1.655397f + Z * 0.036152f;
+  float b =  X * 0.051713f - Y * 0.121364f + Z * 1.011530f;
+
+  if (r > b && r > g && r > 1.0f) {
+    // red is too big
+    g = g / r;
+    b = b / r;
+    r = 1.0f;
+  }
+  else if (g > b && g > r && g > 1.0f) {
+    // green is too big
+    r = r / g;
+    b = b / g;
+    g = 1.0f;
+  }
+  else if (b > r && b > g && b > 1.0f) {
+    // blue is too big
+    r = r / b;
+    g = g / b;
+    b = 1.0f;
+  }
+
+  // Apply gamma correction
+  r = r <= 0.0031308f ? 12.92f * r : (1.0f + 0.055f) * pow(r, (1.0f / 2.4f)) - 0.055f;
+  g = g <= 0.0031308f ? 12.92f * g : (1.0f + 0.055f) * pow(g, (1.0f / 2.4f)) - 0.055f;
+  b = b <= 0.0031308f ? 12.92f * b : (1.0f + 0.055f) * pow(b, (1.0f / 2.4f)) - 0.055f;
+
+  if (r > b && r > g) {
+    // red is biggest
+    if (r > 1.0f) {
+      g = g / r;
+      b = b / r;
+      r = 1.0f;
+    }
+  }
+  else if (g > b && g > r) {
+    // green is biggest
+    if (g > 1.0f) {
+      r = r / g;
+      b = b / g;
+      g = 1.0f;
+    }
+  }
+  else if (b > r && b > g) {
+    // blue is biggest
+    if (b > 1.0f) {
+      r = r / b;
+      g = g / b;
+      b = 1.0f;
+    }
+  }
+
+  r = r < 0 ? 0 : r;
+  g = g < 0 ? 0 : g;
+  b = b < 0 ? 0 : b;
+
+  rgbw[0] = (int) (r * 255.0f); rgbw[1] = (int) (g * 255.0f); rgbw[2] = (int) (b * 255.0f); rgbw[3] = 0;
+}
+
+void convert_ct() {
+  int hectemp = 10000 / ct;
+  int r, g, b;
+  if (hectemp <= 66) {
+    r = 255;
+    g = 99.4708025861 * log(hectemp) - 161.1195681661;
+    b = hectemp <= 19 ? 0 : (138.5177312231 * log(hectemp - 10) - 305.0447927307);
+  } else {
+    r = 329.698727446 * pow(hectemp - 60, -0.1332047592);
+    g = 288.1221695283 * pow(hectemp - 60, -0.0755148492);
+    b = 255;
+  }
+  r = r > 255 ? 255 : r;
+  g = g > 255 ? 255 : g;
+  b = b > 255 ? 255 : b;
+  rgbw[0] = r * (bri / 255.0f); rgbw[1] = g * (bri / 255.0f); rgbw[2] = b * (bri / 255.0f); rgbw[3] = bri;
+}
+
+void handleNotFound() {
+  String message = "File Not Found\n\n";
+  message += "URI: ";
+  message += server.uri();
+  message += "\nMethod: ";
+  message += (server.method() == HTTP_GET) ? "GET" : "POST";
+  message += "\nArguments: ";
+  message += server.args();
+  message += "\n";
+  for (uint8_t i = 0; i < server.args(); i++) {
+    message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
+  }
+  server.send(404, "text/plain", message);
+}
+
+void apply_scene(uint8_t new_scene) {
+  if ( new_scene == 0) {
+    bri = 144; ct = 447; color_mode = 2; convert_ct();
+  } else if ( new_scene == 1) {
+    bri = 254; ct = 346; color_mode = 2; convert_ct();
+  } else if ( new_scene == 2) {
+    bri = 254; ct = 233; color_mode = 2; convert_ct();
+  }  else if ( new_scene == 3) {
+    bri = 254; ct = 156; color_mode = 2; convert_ct();
+  }  else if ( new_scene == 4) {
+    bri = 77; ct = 367; color_mode = 2; convert_ct();
+  }  else if ( new_scene == 5) {
+    bri = 254; ct = 447; color_mode = 2; convert_ct();
+  }  else if ( new_scene == 6) {
+    bri = 1; x = 0.561; y = 0.4042; color_mode = 1; convert_xy();
+  }  else if ( new_scene == 7) {
+    bri = 203; x = 0.380328; y = 0.39986; color_mode = 1; convert_xy();
+  }  else if ( new_scene == 8) {
+    bri = 112; x = 0.359168; y = 0.28807; color_mode = 1; convert_xy();
+  }  else if ( new_scene == 9) {
+    bri = 142; x = 0.267102; y = 0.23755; color_mode = 1; convert_xy();
+  }  else if ( new_scene == 10) {
+    bri = 216; x = 0.393209; y = 0.29961; color_mode = 1; convert_xy();
+  }
+}
+
+
+void setup() {
+  EEPROM.begin(512);
+
+  pins[0] = red_pin;
+  pins[1] = green_pin;
+  pins[2] = blue_pin;
+  pins[3] = white_pin;
+  WiFiManager wifiManager;
+  wifiManager.autoConnect("New Hue Light");
+  analogWriteRange(1024);
+  analogWriteFreq(4096);
+
+  //WiFi.config(strip_ip, gateway_ip, subnet_mask);
+
+  apply_scene(EEPROM.read(2));
+  step_level[0] = rgbw[0] / 350.0f; step_level[1] = rgbw[1] / 350.0f; step_level[2] = rgbw[2] / 350.0f; step_level[3] = rgbw[3] / 350.0f;
+
+  if (EEPROM.read(1) == 1 || (EEPROM.read(1) == 0 && EEPROM.read(0) == 1)) {
+    light_state = true;
+  } else {
+    while (WiFi.status() != WL_CONNECTED) {
+      analogWrite(pins[0], 10);
+      delay(250);
+      analogWrite(pins[0], 0);
+      delay(250);
+    }
+    // Show that we are connected
+    analogWrite(pins[1], 10);
+    delay(500);
+    analogWrite(pins[1], 0);
+  }
+
+  WiFi.macAddress(mac);
+
+  // Port defaults to 8266
+  // ArduinoOTA.setPort(8266);
+
+  // Hostname defaults to esp8266-[ChipID]
+  // ArduinoOTA.setHostname("myesp8266");
+
+  // No authentication by default
+  // ArduinoOTA.setPassword((const char *)"123");
+
+  ArduinoOTA.begin();
+
+  pinMode(LED_BUILTIN, OUTPUT);     // Initialize the LED_BUILTIN pin as an output
+  digitalWrite(LED_BUILTIN, HIGH);  // Turn the LED off by making the voltage HIGH
+
+  server.on("/switch", []() {
+    server.send(200, "text/plain", "OK");
+    int button;
+    for (uint8_t i = 0; i < server.args(); i++) {
+      if (server.argName(i) == "button") {
+        button = server.arg(i).toInt();
+      }
+    }
+    if (button == 1000) {
+      if (light_state == false) {
+        light_state = true;
+        scene = 0;
+      } else {
+        apply_scene(scene);
+        scene++;
+        if (scene == 11) {
+          scene = 0;
+        }
+      }
+    } else if (button == 2000) {
+      if (light_state == false) {
+        bri = 30;
+        light_state = true;
+      } else {
+        bri += 30;
+      }
+      if (bri > 255) bri = 255;
+      if (color_mode == 1) convert_xy();
+      else if (color_mode == 2) convert_ct();
+      else if (color_mode == 3) convert_hue();
+    } else if (button == 3000 && light_state == true) {
+      bri -= 30;
+      if (bri < 1) bri = 1;
+      else {
+        if (color_mode == 1) convert_xy();
+        else if (color_mode == 2) convert_ct();
+        else if (color_mode == 3) convert_hue();
+      }
+    } else if (button == 4000) {
+      light_state = false;
+    }
+    for (uint8_t color = 0; color < pwm_channels; color++) {
+      if (light_state) {
+        step_level[color] = (rgbw[color] - current_rgbw[color]) / 54;
+      } else {
+        step_level[color] = current_rgbw[color] / 54;
+      }
+    }
+  });
+
+
+  server.on("/set", []() {
+    light_state = true;
+    float transitiontime = 4;
+    for (uint8_t i = 0; i < server.args(); i++) {
+      if (server.argName(i) == "on") {
+        if (server.arg(i) == "True" || server.arg(i) == "true") {
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) != 1) {
+            EEPROM.write(0, 1);
+            EEPROM.commit();
+          }
+          light_state = true;
+        }
+        else {
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) != 0) {
+            EEPROM.write(0, 0);
+            EEPROM.commit();
+          }
+          light_state = false;
+        }
+      }
+      else if (server.argName(i) == "r") {
+        rgbw[0] = server.arg(i).toInt();
+        color_mode = 0;
+      }
+      else if (server.argName(i) == "g") {
+        rgbw[1] = server.arg(i).toInt();
+        color_mode = 0;
+      }
+      else if (server.argName(i) == "b") {
+        rgbw[2] = server.arg(i).toInt();
+        color_mode = 0;
+      }
+      else if (server.argName(i) == "w") {
+        rgbw[3] = server.arg(i).toInt();
+        color_mode = 0;
+      }
+      else if (server.argName(i) == "x") {
+        x = server.arg(i).toFloat();
+        color_mode = 1;
+      }
+      else if (server.argName(i) == "y") {
+        y = server.arg(i).toFloat();
+        color_mode = 1;
+      }
+      else if (server.argName(i) == "bri") {
+        if (server.arg(i).toInt() != 0)
+          bri = server.arg(i).toInt();
+      }
+      else if (server.argName(i) == "bri_inc") {
+        bri += server.arg(i).toInt();
+        if (bri > 255) bri = 255;
+        else if (bri < 0) bri = 0;
+      }
+      else if (server.argName(i) == "ct") {
+        ct = server.arg(i).toInt();
+        color_mode = 2;
+      }
+      else if (server.argName(i) == "sat") {
+        sat = server.arg(i).toInt();
+        color_mode = 3;
+      }
+      else if (server.argName(i) == "hue") {
+        hue = server.arg(i).toInt();
+        color_mode = 3;
+      }
+      else if (server.argName(i) == "alert") {
+        if (light_state) {
+          current_rgbw[0] = 0; current_rgbw[1] = 0; current_rgbw[2] = 0; current_rgbw[3] = 0;
+        } else {
+          current_rgbw[0] = 255; current_rgbw[1] = 255; current_rgbw[2] = 255; current_rgbw[3] = 255;
+        }
+      }
+      else if (server.argName(i) == "transitiontime") {
+        transitiontime = server.arg(i).toInt();
+      }
+    }
+    server.send(200, "text/plain", "OK, x: " + (String)x + ", y:" + (String)y + ", bri:" + (String)bri + ", ct:" + ct + ", colormode:" + color_mode + ", state:" + light_state);
+    if (color_mode == 1 && light_state == true) {
+      convert_xy();
+    } else if (color_mode == 2 && light_state == true) {
+      convert_ct();
+    } else if (color_mode == 3 && light_state == true) {
+      convert_hue();
+    }
+    transitiontime *= 60.0;
+    for (uint8_t color = 0; color < pwm_channels; color++) {
+      if (light_state) {
+        step_level[color] = (rgbw[color] - current_rgbw[color]) / transitiontime;
+      } else {
+        step_level[color] = current_rgbw[color] / transitiontime;
+      }
+    }
+  });
+
+  server.on("/get", []() {
+    server.send(200, "text/plain", "{\"R\":" + (String)rgbw[0] + ", \"G\": " + (String)rgbw[1] + ", \"B\":" + (String)rgbw[2] + ", \"W\":" + (String)rgbw[3] + ", \"bri\":" + (String)bri + ", \"xy\": [" + (String)x + "," + (String)y + "], \"ct\":" + (String)ct + ", \"sat\": " + (String)sat + ", \"hue\": " + (String)hue + ", \"colormode\":" + color_mode + "}");
+  });
+
+  server.on("/detect", []() {
+    server.send(200, "text/plain", "{\"hue\": \"bulb\",\"lights\": 1,\"type\": \"rgb\",\"mac\": \"" + String(mac[5], HEX) + ":"  + String(mac[4], HEX) + ":" + String(mac[3], HEX) + ":" + String(mac[2], HEX) + ":" + String(mac[1], HEX) + ":" + String(mac[0], HEX) + "\"}");
+  });
+
+  server.on("/", []() {
+    float transitiontime = 200;
+    if (server.hasArg("startup")) {
+      if (  EEPROM.read(1) != server.arg("startup").toInt()) {
+        EEPROM.write(1, server.arg("startup").toInt());
+        EEPROM.commit();
+      }
+    }
+
+    if (server.hasArg("scene")) {
+      if (server.arg("bri") == "" && server.arg("hue") == "" && server.arg("ct") == "" && server.arg("sat") == "") {
+        if (  EEPROM.read(2) != server.arg("scene").toInt() && EEPROM.read(1) < 2) {
+          EEPROM.write(2, server.arg("scene").toInt());
+          EEPROM.commit();
+        }
+        apply_scene(server.arg("scene").toInt());
+      } else {
+        if (server.arg("bri") != "") {
+          bri = server.arg("bri").toInt();
+        }
+        if (server.arg("hue") != "") {
+          hue = server.arg("hue").toInt();
+        }
+        if (server.arg("sat") != "") {
+          sat = server.arg("sat").toInt();
+        }
+        if (server.arg("ct") != "") {
+          ct = server.arg("ct").toInt();
+        }
+        if (server.arg("colormode") == "1" && light_state == true) {
+          convert_xy();
+        } else if (server.arg("colormode") == "2" && light_state == true) {
+          convert_ct();
+        } else if (server.arg("colormode") == "3" && light_state == true) {
+          convert_hue();
+        }
+        color_mode = server.arg("colormode").toInt();
+      }
+    } else if (server.hasArg("on")) {
+      if (server.arg("on") == "true") {
+        light_state = true; {
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) != 1) {
+            EEPROM.write(0, 1);
+          }
+        }
+      } else {
+        light_state = false;
+        if (EEPROM.read(1) == 0 && EEPROM.read(0) != 0) {
+          EEPROM.write(0, 0);
+        }
+      }
+      EEPROM.commit();
+    } else if (server.hasArg("alert")) {
+      if (light_state) {
+        current_rgbw[0] = 0; current_rgbw[1] = 0; current_rgbw[2] = 0; current_rgbw[3] = 0;
+      } else {
+        current_rgbw[3] = 255;
+      }
+    }
+    for (uint8_t color = 0; color < pwm_channels; color++) {
+      if (light_state) {
+        step_level[color] = ((float)rgbw[color] - current_rgbw[color]) / transitiontime;
+      } else {
+        step_level[color] = current_rgbw[color] / transitiontime;
+      }
+    }
+    if (server.hasArg("reset")) {
+      ESP.reset();
+    }
+
+
+    String http_content = "<!doctype html>";
+    http_content += "<html>";
+    http_content += "<head>";
+    http_content += "<meta charset=\"utf-8\">";
+    http_content += "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">";
+    http_content += "<title>Light Setup</title>";
+    http_content += "<link rel=\"stylesheet\" href=\"https://unpkg.com/purecss@0.6.2/build/pure-min.css\">";
+    http_content += "</head>";
+    http_content += "<body>";
+    http_content += "<fieldset>";
+    http_content += "<h3>Light Setup</h3>";
+    http_content += "<form class=\"pure-form pure-form-aligned\" action=\"/\" method=\"post\">";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"power\"><strong>Power</strong></label>";
+    http_content += "<a class=\"pure-button"; if (light_state) http_content += "  pure-button-primary"; http_content += "\" href=\"/?on=true\">ON</a>";
+    http_content += "<a class=\"pure-button"; if (!light_state) http_content += "  pure-button-primary"; http_content += "\" href=\"/?on=false\">OFF</a>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"startup\">Startup</label>";
+    http_content += "<select onchange=\"this.form.submit()\" id=\"startup\" name=\"startup\">";
+    http_content += "<option "; if (EEPROM.read(1) == 0) http_content += "selected=\"selected\""; http_content += " value=\"0\">Last state</option>";
+    http_content += "<option "; if (EEPROM.read(1) == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">On</option>";
+    http_content += "<option "; if (EEPROM.read(1) == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">Off</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"scene\">Scene</label>";
+    http_content += "<select onchange = \"this.form.submit()\" id=\"scene\" name=\"scene\">";
+    http_content += "<option "; if (EEPROM.read(2) == 0) http_content += "selected=\"selected\""; http_content += " value=\"0\">Relax</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">Read</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">Concentrate</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 3) http_content += "selected=\"selected\""; http_content += " value=\"3\">Energize</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 4) http_content += "selected=\"selected\""; http_content += " value=\"4\">Bright</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 5) http_content += "selected=\"selected\""; http_content += " value=\"5\">Dimmed</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 6) http_content += "selected=\"selected\""; http_content += " value=\"6\">Nightlight</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 7) http_content += "selected=\"selected\""; http_content += " value=\"7\">Savanna sunset</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 8) http_content += "selected=\"selected\""; http_content += " value=\"8\">Tropical twilight</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 9) http_content += "selected=\"selected\""; http_content += " value=\"9\">Arctic aurora</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 10) http_content += "selected=\"selected\""; http_content += " value=\"10\">Spring blossom</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<br>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"state\"><strong>State</strong></label>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"bri\">Bri</label>";
+    http_content += "<input id=\"bri\" name=\"bri\" type=\"text\" placeholder=\"" + (String)bri + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"hue\">Hue</label>";
+    http_content += "<input id=\"hue\" name=\"hue\" type=\"text\" placeholder=\"" + (String)hue + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"sat\">Sat</label>";
+    http_content += "<input id=\"sat\" name=\"sat\" type=\"text\" placeholder=\"" + (String)sat + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"ct\">CT</label>";
+    http_content += "<input id=\"ct\" name=\"ct\" type=\"text\" placeholder=\"" + (String)ct + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"colormode\">Color</label>";
+    http_content += "<select id=\"colormode\" name=\"colormode\">";
+    http_content += "<option "; if (color_mode == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">xy</option>";
+    http_content += "<option "; if (color_mode == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">ct</option>";
+    http_content += "<option "; if (color_mode == 3) http_content += "selected=\"selected\""; http_content += " value=\"3\">hue</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-controls\">";
+    http_content += "<span class=\"pure-form-message\"><a href=\"/?alert=1\">alert</a> or <a href=\"/?reset=1\">reset</a></span>";
+    http_content += "<label for=\"cb\" class=\"pure-checkbox\">";
+    http_content += "</label>";
+    http_content += "<button type=\"submit\" class=\"pure-button pure-button-primary\">Save</button>";
+    http_content += "</div>";
+    http_content += "</fieldset>";
+    http_content += "</form>";
+    http_content += "</body>";
+    http_content += "</html>";
+
+
+    server.send(200, "text/html", http_content);
+
+  });
+
+
+  server.onNotFound(handleNotFound);
+
+  server.begin();
+}
+
+void lightEngine() {
+  for (uint8_t color = 0; color < pwm_channels; color++) {
+    if (light_state) {
+      if (rgbw[color] != current_rgbw[color] ) {
+        current_rgbw[color] += step_level[color];
+        if ((step_level[color] > 0.0f && current_rgbw[color] > rgbw[color]) || (step_level[color] < 0.0f && current_rgbw[color] < rgbw[color])) current_rgbw[color] = rgbw[color];
+        analogWrite(pins[color], (int)(current_rgbw[color] * 4));
+      }
+    } else {
+      if (current_rgbw[color] != 0) {
+        current_rgbw[color] -= step_level[color];
+        if (current_rgbw[color] < 0.0f) current_rgbw[color] = 0;
+        analogWrite(pins[color], (int)(current_rgbw[color] * 4));
+      }
+    }
+  }
+  delay(1);
+}
+
+void loop() {
+  ArduinoOTA.handle();
+  server.handleClient();
+  lightEngine();
+}
