--- Lights/PlatformIO/RgbwHueStrip/.gitignore
@@ -1,4 +1,0 @@
-.pioenvs
-.piolibdeps
-.clang_complete
-.gcc-flags.json
