+++ Lights/PlatformIO/SK6812HueStrip/.travis.yml
