+++ Lights/PlatformIO/SK6812HueStrip/platformio.ini
@@ -11,6 +11,7 @@
 [common]
 lib_deps_external =
   NeoPixelBus
+  WiFiManager
 
 [env:nodemcuv2]
 platform = espressif8266
