+++ Lights/PlatformIO/SK6812HueStrip/src/SK6812HueStrip.cpp
@@ -4,27 +4,20 @@
 #include <ESP8266mDNS.h>
 #include <WiFiUdp.h>
 #include <ArduinoOTA.h>
-#include <NeoPixelBus.h>
-
 #include <ESP8266WebServer.h>
-
-
-const char* ssid = "MikroTik";
-const char* password = "nustiuceparola";
+#include <NeoPixelBus.h>
+#include <WiFiManager.h>
+#include <EEPROM.h>
 
 #define lightsCount 3
 #define pixelCount 30
-
-// Available scenes: 0 = Relax, 1 = Read, 2 = Concentrate, 3 = Energize, 4 = Dimmed, 5 = Bright, 6 = Night
-#define default_scene 0
-#define startup_on true
 
 // if you want to setup static ip uncomment these 3 lines and line 72
 //IPAddress strip_ip ( 192,  168,   10,  95);
 //IPAddress gateway_ip ( 192,  168,   10,   1);
 //IPAddress subnet_mask(255, 255, 255,   0);
 
-uint8_t rgbw[lightsCount][4], color_mode[lightsCount], scene = default_scene;
+uint8_t rgbw[lightsCount][4], color_mode[lightsCount], scene;
 bool light_state[lightsCount];
 int transitiontime[lightsCount], ct[lightsCount], hue[lightsCount], bri[lightsCount], sat[lightsCount];
 float step_level[lightsCount][4], current_rgbw[lightsCount][4], x[lightsCount], y[lightsCount];
@@ -215,6 +208,7 @@ void infoLight(RgbwColor color) {
   }
 }
 
+
 void apply_scene(uint8_t new_scene, uint8_t light) {
   if ( new_scene == 0) {
     bri[light] = 144; ct[light] = 447; color_mode[light] = 2; convert_ct(light);
@@ -229,7 +223,7 @@ void apply_scene(uint8_t new_scene, uint8_t light) {
   }  else if ( new_scene == 5) {
     bri[light] = 254; ct[light] = 447; color_mode[light] = 2; convert_ct(light);
   }  else if ( new_scene == 6) {
-    bri[light] = 1; x[light] = 0, 561; y[light] = 0, 4042; color_mode[light] = 1; convert_xy(light);
+    bri[light] = 1; x[light] = 0.561; y[light] = 0.4042; color_mode[light] = 1; convert_xy(light);
   }  else if ( new_scene == 7) {
     bri[light] = 203; x[light] = 0.380328; y[light] = 0.39986; color_mode[light] = 1; convert_xy(light);
   }  else if ( new_scene == 8) {
@@ -245,21 +239,19 @@ void apply_scene(uint8_t new_scene, uint8_t light) {
 void setup() {
   strip.Begin();
   strip.Show();
+  EEPROM.begin(512);
 
-  // Show that the NeoPixels are alive
-  delay(120); // Apparently needed to make the first few pixels animate correctly
+  WiFiManager wifiManager;
+  wifiManager.autoConnect("New Hue Light");
 
   //WiFi.config(strip_ip, gateway_ip, subnet_mask);
 
-  WiFi.mode(WIFI_STA);
-  WiFi.begin(ssid, password);
-
   for (int i = 0; i < lightsCount; i++) {
-    apply_scene(default_scene, i);
+    apply_scene(EEPROM.read(2), i);
     step_level[i][0] = rgbw[i][0] / 350.0f; step_level[i][1] = rgbw[i][1] / 350.0f; step_level[i][2] = rgbw[i][2] / 350.0f; step_level[i][3] = rgbw[i][3] / 350.0f;
   }
 
-  if (startup_on == true) {
+  if (EEPROM.read(1) == 1 || (EEPROM.read(1) == 0 && EEPROM.read(0) == 1)) {
     for (int i = 0; i < lightsCount; i++) {
       light_state[i] = true;
     }
@@ -271,11 +263,6 @@ void setup() {
     }
     // Show that we are connected
     infoLight(green);
-
-    //setup default warm_white on power on
-    for (int i = 0; i < lightsCount; i++) {
-      rgbw[i][0] = 254; rgbw[i][1] = 145; rgbw[i][2] = 40;
-    }
 
   }
 
@@ -298,6 +285,7 @@ void setup() {
 
   server.on("/switch", []() {
     server.send(200, "text/plain", "OK");
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
     int button;
     for (uint8_t i = 0; i < server.args(); i++) {
       if (server.argName(i) == "button") {
@@ -338,16 +326,12 @@ void setup() {
       } else if (button == 4000) {
         light_state[i] = false;
       }
-      if (light_state[i]) {
-        step_level[i][0] = (rgbw[i][0] - current_rgbw[i][0]) / 54;
-        step_level[i][1] = (rgbw[i][1] - current_rgbw[i][1]) / 54;
-        step_level[i][2] = (rgbw[i][2] - current_rgbw[i][2]) / 54;
-        step_level[i][3] = (rgbw[i][3] - current_rgbw[i][3]) / 54;
-      } else {
-        step_level[i][0] = current_rgbw[i][0] / 54;
-        step_level[i][1] = current_rgbw[i][1] / 54;
-        step_level[i][2] = current_rgbw[i][2] / 54;
-        step_level[i][3] = current_rgbw[i][3] / 54;
+      for (uint8_t j = 0; j <= 3; j++) {
+        if (light_state[i]) {
+          step_level[i][j] = ((float)rgbw[i][j] - current_rgbw[i][j]) / transitiontime;
+        } else {
+          step_level[i][j] = current_rgbw[i][j] / transitiontime;
+        }
       }
     }
   });
@@ -358,15 +342,21 @@ void setup() {
     for (uint8_t i = 0; i < server.args(); i++) {
       if (server.argName(i) == "light") {
         light = server.arg(i).toInt() - 1;
-        light_state[light] = true;
       }
       else if (server.argName(i) == "on") {
-        if (server.arg(i) == "True") {
+        if (server.arg(i) == "True" || server.arg(i) == "true") {
           light_state[light] = true;
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) == 0) {
+            EEPROM.write(0, 1);
+          }
         }
         else {
           light_state[light] = false;
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) == 1) {
+            EEPROM.write(0, 0);
+          }
         }
+        EEPROM.commit();
       }
       else if (server.argName(i) == "r") {
         rgbw[light][0] = server.arg(i).toInt();
@@ -393,6 +383,7 @@ void setup() {
         color_mode[light] = 1;
       }
       else if (server.argName(i) == "bri") {
+        light_state[light] = true;
         if (server.arg(i).toInt() != 0)
           bri[light] = server.arg(i).toInt();
       }
@@ -413,11 +404,18 @@ void setup() {
         hue[light] = server.arg(i).toInt();
         color_mode[light] = 3;
       }
+      else if (server.argName(i) == "alert") {
+        if (light_state[light]) {
+          current_rgbw[light][0] = 0; current_rgbw[light][1] = 0; current_rgbw[light][2] = 0; current_rgbw[light][3] = 0;
+        } else {
+          current_rgbw[light][3] = 255;
+        }
+      }
       else if (server.argName(i) == "transitiontime") {
         transitiontime = server.arg(i).toInt();
       }
     }
-    transitiontime *= 10;
+    transitiontime *= 10 - (pixelCount / 40); //every extra led add a small delay that need to be counted
     server.send(200, "text/plain", "OK, x: " + (String)x[light] + ", y:" + (String)y[light] + ", bri:" + (String)bri[light] + ", ct:" + ct[light] + ", colormode:" + color_mode[light] + ", state:" + light_state[light]);
     if (color_mode[light] == 1 && light_state[light] == true) {
       convert_xy(light);
@@ -426,16 +424,12 @@ void setup() {
     } else if (color_mode[light] == 3 && light_state[light] == true) {
       convert_hue(light);
     }
-    if (light_state[light]) {
-      step_level[light][0] = ((float)rgbw[light][0] - current_rgbw[light][0]) / transitiontime;
-      step_level[light][1] = ((float)rgbw[light][1] - current_rgbw[light][1]) / transitiontime;
-      step_level[light][2] = ((float)rgbw[light][2] - current_rgbw[light][2]) / transitiontime;
-      step_level[light][3] = ((float)rgbw[light][3] - current_rgbw[light][3]) / transitiontime;
-    } else {
-      step_level[light][0] = current_rgbw[light][0] / transitiontime;
-      step_level[light][1] = current_rgbw[light][1] / transitiontime;
-      step_level[light][2] = current_rgbw[light][2] / transitiontime;
-      step_level[light][3] = current_rgbw[light][3] / transitiontime;
+    for (uint8_t j = 0; j <= 3; j++) {
+      if (light_state[light]) {
+        step_level[light][j] = ((float)rgbw[light][j] - current_rgbw[light][j]) / transitiontime;
+      } else {
+        step_level[light][j] = current_rgbw[light][j] / transitiontime;
+      }
     }
   });
 
@@ -453,9 +447,162 @@ void setup() {
     server.send(200, "text/plain", "{\"hue\": \"strip\",\"lights\": " + (String)lightsCount + ",\"type\": \"rgb\",\"mac\": \"" + String(mac[5], HEX) + ":"  + String(mac[4], HEX) + ":" + String(mac[3], HEX) + ":" + String(mac[2], HEX) + ":" + String(mac[1], HEX) + ":" + String(mac[0], HEX) + "\"}");
   });
 
-  server.on("/reset", []() {
-    server.send(200, "text/plain", "reset");
-    ESP.reset();
+  server.on("/", []() {
+    float transitiontime = (10 - (pixelCount / 40)) * 4;
+    if (server.hasArg("startup")) {
+      if (  EEPROM.read(1) != server.arg("startup").toInt()) {
+        EEPROM.write(1, server.arg("startup").toInt());
+        EEPROM.commit();
+      }
+    }
+
+    for (int light = 0; light < lightsCount; light++) {
+      if (server.hasArg("scene")) {
+        if (server.arg("bri") == "" && server.arg("hue") == "" && server.arg("ct") == "" && server.arg("sat") == "") {
+          if (  EEPROM.read(2) != server.arg("scene").toInt()) {
+            EEPROM.write(2, server.arg("scene").toInt());
+            EEPROM.commit();
+          }
+          apply_scene(server.arg("scene").toInt(), light);
+        } else {
+          if (server.arg("bri") != "") {
+            bri[light] = server.arg("bri").toInt();
+          }
+          if (server.arg("hue") != "") {
+            hue[light] = server.arg("hue").toInt();
+          }
+          if (server.arg("sat") != "") {
+            sat[light] = server.arg("sat").toInt();
+          }
+          if (server.arg("ct") != "") {
+            ct[light] = server.arg("ct").toInt();
+          }
+          if (server.arg("colormode") == "1" && light_state[light] == true) {
+            convert_xy(light);
+          } else if (server.arg("colormode") == "2" && light_state[light] == true) {
+            convert_ct(light);
+          } else if (server.arg("colormode") == "3" && light_state[light] == true) {
+            convert_hue(light);
+          }
+          color_mode[light] = server.arg("colormode").toInt();
+        }
+      } else if (server.hasArg("on")) {
+        if (server.arg("on") == "true") {
+          light_state[light] = true; {
+            if (EEPROM.read(1) == 0 && EEPROM.read(0) == 0) {
+              EEPROM.write(0, 1);
+            }
+          }
+        } else {
+          light_state[light] = false;
+          if (EEPROM.read(1) == 0 && EEPROM.read(0) == 1) {
+            EEPROM.write(0, 0);
+          }
+        }
+        EEPROM.commit();
+      } else if (server.hasArg("alert")) {
+        if (light_state[light]) {
+          current_rgbw[light][0] = 0; current_rgbw[light][1] = 0; current_rgbw[light][2] = 0; current_rgbw[light][3] = 0;
+        } else {
+          current_rgbw[light][3] = 255;
+        }
+      }
+      for (uint8_t j = 0; j <= 3; j++) {
+        if (light_state[light]) {
+          step_level[light][j] = ((float)rgbw[light][j] - current_rgbw[light][j]) / transitiontime;
+        } else {
+          step_level[light][j] = current_rgbw[light][j] / transitiontime;
+        }
+      }
+    }
+    if (server.hasArg("reset")) {
+      ESP.reset();
+    }
+
+
+    String http_content = "<!doctype html>";
+    http_content += "<html>";
+    http_content += "<head>";
+    http_content += "<meta charset=\"utf-8\">";
+    http_content += "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">";
+    http_content += "<title>Light Setup</title>";
+    http_content += "<link rel=\"stylesheet\" href=\"https://unpkg.com/purecss@0.6.2/build/pure-min.css\">";
+    http_content += "</head>";
+    http_content += "<body>";
+    http_content += "<fieldset>";
+    http_content += "<h3>Light Setup</h3>";
+    http_content += "<form class=\"pure-form pure-form-aligned\" action=\"/\" method=\"post\">";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"power\"><strong>Power</strong></label>";
+    http_content += "<a class=\"pure-button"; if (light_state[0]) http_content += "  pure-button-primary"; http_content += "\" href=\"/?on=true\">ON</a>";
+    http_content += "<a class=\"pure-button"; if (!light_state[0]) http_content += "  pure-button-primary"; http_content += "\" href=\"/?on=false\">OFF</a>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"startup\">Startup</label>";
+    http_content += "<select onchange=\"this.form.submit()\" id=\"startup\" name=\"startup\">";
+    http_content += "<option "; if (EEPROM.read(1) == 0) http_content += "selected=\"selected\""; http_content += " value=\"0\">Last state</option>";
+    http_content += "<option "; if (EEPROM.read(1) == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">On</option>";
+    http_content += "<option "; if (EEPROM.read(1) == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">Off</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"scene\">Scene</label>";
+    http_content += "<select onchange = \"this.form.submit()\" id=\"scene\" name=\"scene\">";
+    http_content += "<option "; if (EEPROM.read(2) == 0) http_content += "selected=\"selected\""; http_content += " value=\"0\">Relax</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">Read</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">Concentrate</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 3) http_content += "selected=\"selected\""; http_content += " value=\"3\">Energize</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 4) http_content += "selected=\"selected\""; http_content += " value=\"4\">Bright</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 5) http_content += "selected=\"selected\""; http_content += " value=\"5\">Dimmed</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 6) http_content += "selected=\"selected\""; http_content += " value=\"6\">Nightlight</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 7) http_content += "selected=\"selected\""; http_content += " value=\"7\">Savanna sunset</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 8) http_content += "selected=\"selected\""; http_content += " value=\"8\">Tropical twilight</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 9) http_content += "selected=\"selected\""; http_content += " value=\"9\">Arctic aurora</option>";
+    http_content += "<option "; if (EEPROM.read(2) == 10) http_content += "selected=\"selected\""; http_content += " value=\"10\">Spring blossom</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<br>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"state\"><strong>State</strong></label>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"bri\">Bri</label>";
+    http_content += "<input id=\"bri\" name=\"bri\" type=\"text\" placeholder=\"" + (String)bri[0] + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"hue\">Hue</label>";
+    http_content += "<input id=\"hue\" name=\"hue\" type=\"text\" placeholder=\"" + (String)hue[0] + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"sat\">Sat</label>";
+    http_content += "<input id=\"sat\" name=\"sat\" type=\"text\" placeholder=\"" + (String)sat[0] + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"ct\">CT</label>";
+    http_content += "<input id=\"ct\" name=\"ct\" type=\"text\" placeholder=\"" + (String)ct[0] + "\">";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-control-group\">";
+    http_content += "<label for=\"colormode\">Color</label>";
+    http_content += "<select id=\"colormode\" name=\"colormode\">";
+    http_content += "<option "; if (color_mode[0] == 1) http_content += "selected=\"selected\""; http_content += " value=\"1\">xy</option>";
+    http_content += "<option "; if (color_mode[0] == 2) http_content += "selected=\"selected\""; http_content += " value=\"2\">ct</option>";
+    http_content += "<option "; if (color_mode[0] == 3) http_content += "selected=\"selected\""; http_content += " value=\"3\">hue</option>";
+    http_content += "</select>";
+    http_content += "</div>";
+    http_content += "<div class=\"pure-controls\">";
+    http_content += "<span class=\"pure-form-message\"><a href=\"/?alert=1\">alert</a> or <a href=\"/?reset=1\">reset</a></span>";
+    http_content += "<label for=\"cb\" class=\"pure-checkbox\">";
+    http_content += "</label>";
+    http_content += "<button type=\"submit\" class=\"pure-button pure-button-primary\">Save</button>";
+    http_content += "</div>";
+    http_content += "</fieldset>";
+    http_content += "</form>";
+    http_content += "</body>";
+    http_content += "</html>";
+
+
+    server.send(200, "text/html", http_content);
+
   });
 
   server.onNotFound(handleNotFound);
@@ -467,14 +614,10 @@ void lightEngine() {
   for (int i = 0; i < lightsCount; i++) {
     if (light_state[i]) {
       if (rgbw[i][0] != current_rgbw[i][0] || rgbw[i][1] != current_rgbw[i][1] || rgbw[i][2] != current_rgbw[i][2] || rgbw[i][3] != current_rgbw[i][3]) {
-        if (rgbw[i][0] != current_rgbw[i][0]) current_rgbw[i][0] += step_level[i][0];
-        if (rgbw[i][1] != current_rgbw[i][1]) current_rgbw[i][1] += step_level[i][1];
-        if (rgbw[i][2] != current_rgbw[i][2]) current_rgbw[i][2] += step_level[i][2];
-        if (rgbw[i][3] != current_rgbw[i][3]) current_rgbw[i][3] += step_level[i][3];
-        if ((step_level[i][0] > 0.0 && current_rgbw[i][0] > rgbw[i][0]) || (step_level[i][0] < 0.0 && current_rgbw[i][0] < rgbw[i][0])) current_rgbw[i][0] = rgbw[i][0];
-        if ((step_level[i][1] > 0.0 && current_rgbw[i][1] > rgbw[i][1]) || (step_level[i][1] < 0.0 && current_rgbw[i][1] < rgbw[i][1])) current_rgbw[i][1] = rgbw[i][1];
-        if ((step_level[i][2] > 0.0 && current_rgbw[i][2] > rgbw[i][2]) || (step_level[i][2] < 0.0 && current_rgbw[i][2] < rgbw[i][2])) current_rgbw[i][2] = rgbw[i][2];
-        if ((step_level[i][3] > 0.0 && current_rgbw[i][3] > rgbw[i][3]) || (step_level[i][3] < 0.0 && current_rgbw[i][3] < rgbw[i][3])) current_rgbw[i][3] = rgbw[i][3];
+        for (uint8_t k = 0; k <= 3; k++) {
+          if (rgbw[i][k] != current_rgbw[i][k]) current_rgbw[i][k] += step_level[i][k];
+          if ((step_level[i][k] > 0.0 && current_rgbw[i][k] > rgbw[i][k]) || (step_level[i][k] < 0.0 && current_rgbw[i][k] < rgbw[i][k])) current_rgbw[i][k] = rgbw[i][k];
+        }
         for (int j = 0; j < pixelCount / lightsCount ; j++)
         {
           strip.SetPixelColor(j + i * pixelCount / lightsCount, RgbwColor((int)current_rgbw[i][0], (int)current_rgbw[i][1], (int)current_rgbw[i][2], (int)current_rgbw[i][3]));
@@ -483,14 +626,10 @@ void lightEngine() {
       }
     } else {
       if (current_rgbw[i][0] != 0 || current_rgbw[i][1] != 0 || current_rgbw[i][2] != 0 || current_rgbw[i][3] != 0) {
-        if (current_rgbw[i][0] != 0) current_rgbw[i][0] -= step_level[i][0];
-        if (current_rgbw[i][1] != 0) current_rgbw[i][1] -= step_level[i][1];
-        if (current_rgbw[i][2] != 0) current_rgbw[i][2] -= step_level[i][2];
-        if (current_rgbw[i][3] != 0) current_rgbw[i][3] -= step_level[i][3];
-        if (current_rgbw[i][0] < 0) current_rgbw[i][0] = 0;
-        if (current_rgbw[i][1] < 0) current_rgbw[i][1] = 0;
-        if (current_rgbw[i][2] < 0) current_rgbw[i][2] = 0;
-        if (current_rgbw[i][3] < 0) current_rgbw[i][3] = 0;
+        for (uint8_t k = 0; k <= 3; k++) {
+          if (current_rgbw[i][k] != 0) current_rgbw[i][k] -= step_level[i][k];
+          if (current_rgbw[i][k] < 0) current_rgbw[i][k] = 0;
+        }
         for (int j = 0; j < pixelCount / lightsCount ; j++)
         {
           strip.SetPixelColor(j + i * pixelCount / lightsCount, RgbwColor((int)current_rgbw[i][0], (int)current_rgbw[i][1], (int)current_rgbw[i][2], (int)current_rgbw[i][3]));
@@ -507,3 +646,5 @@ void loop() {
   server.handleClient();
   lightEngine();
 }
+
+
