+++ Lights/PlatformIO/WS2812BHueStrip/.travis.yml
